# docs/swagger.yaml
basePath: /
definitions:
...
  history.PlayerAnswer:
    properties:
      id:
        type: string
      isCorrect:
        type: boolean
      questionIndex:
        type: integer
      roomHistoryId:
        type: string
      roomPlayerId:
        type: string
      selectedIndex:
        description: -1 se não respondeu?
        type: integer
    type: object
  history.PlayerStats:
    properties:
      correctCount:
        type: integer
      id:
        type: string
      nickname:
        type: string
      playerRuntimeId:
        type: string
      roomHistoryId:
        type: string
      score:
        type: integer
      wrongCount:
        type: integer
    type: object
  history.QuestionStats:
    properties:
      correctCount:
        type: integer
      correctIndex:
        type: integer
      countA:
        type: integer
      countB:
        type: integer
      countC:
        type: integer
      countD:
        type: integer
      id:
        type: string
      promptSnapshot:
        type: string
      questionId:
        type: string
      questionIndex:
        type: integer
      roomHistoryId:
        type: string
    type: object
  history.RoomHistory:
    properties:
      answers:
        items:
          $ref: '#/definitions/history.PlayerAnswer'
        type: array
      createdAt:
        type: string
      finishedAt:
        type: string
      id:
        type: string
      players:
        items:
          $ref: '#/definitions/history.PlayerStats'
        type: array
      questions:
        items:
          $ref: '#/definitions/history.QuestionStats'
        type: array
      quizId:
        type: string
      quizTitleSnapshot:
        type: string
      roomId:
        type: string
      startedAt:
        type: string
      status:
        type: string
      teacherId:
        type: string
      totalQuestions:
        type: integer
    type: object
  quiz.Question:
    properties:
      correctIndex:
        description: 0..3
        type: integer
      createdAt:
        type: string
      id:
        type: string
      optionA:
        description: "0"
        type: string
      optionB:
        description: "1"
        type: string
      optionC:
        description: "2"
        type: string
      optionD:
        description: "3"
        type: string
      prompt:
        description: Enunciado
        type: string
      quizId:
        type: string
      sortOrder:
        description: Ordem na lista
        type: integer
      updatedAt:
        type: string
    type: object
  quiz.Quiz:
    properties:
      createdAt:
        type: string
      description:
        type: string
      grade:
        description: 'Série (ex: 7º Ano)'
        type: string
      id:
        type: string
      questions:
        items:
          $ref: '#/definitions/quiz.Question'
        type: array
      status:
        description: DRAFT | PUBLISHED
        type: string
      subject:
        description: 'Disciplina (ex: História)'
        type: string
      teacherId:
        description: Owner
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  usecases.CreateQuizInput:
    properties:
      description:
        type: string
      grade:
        type: string
      subject:
        type: string
      teacherID:
        type: string
      title:
        type: string
    type: object
...
  usecases.UpdateQuizInput:
    properties:
      description:
        type: string
      grade:
        type: string
      quizID:
        type: string
      subject:
        type: string
      teacherID:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Suporte RankIt
  description: Backend para a plataforma RankIt (Gamificação em sala de aula).
  title: RankIt API
  version: "1.0"
paths:
...
  /quizzes:
    get:
      description: Retorna todos os quizzes do professor logado.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/quiz.Quiz'
            type: array
      security:
      - BearerAuth: []
      summary: Lista quizzes do professor
      tags:
      - Quizzes
    post:
      consumes:
      - application/json
      description: Cria um quiz vinculado ao professor logado. Status inicial DRAFT.
      parameters:
      - description: Dados do Quiz
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.CreateQuizInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "400":
          description: Erro de validação
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria um novo quiz (Rascunho)
      tags:
      - Quizzes
  /quizzes/{id}:
    delete:
      description: Remove quiz e perguntas. Apenas se DRAFT.
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Não pode deletar quiz publicado
      security:
      - BearerAuth: []
      summary: Remove um quiz
      tags:
      - Quizzes
    get:
      description: Retorna dados do quiz e suas perguntas.
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "404":
          description: Não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detalha um quiz
      tags:
      - Quizzes
    put:
      consumes:
      - application/json
      description: Atualiza título, descrição, etc. Apenas se DRAFT.
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      - description: Dados novos
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.UpdateQuizInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "400":
          description: Erro validação ou estao já publicado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza um quiz (Rascunho)
      tags:
      - Quizzes
  /quizzes/{id}/publish:
    post:
      description: Altera status para PUBLISHED. Valida perguntas.
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Quiz'
        "409":
          description: Quiz inválido para publicação
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Publica um quiz
      tags:
      - Quizzes
...
  /reports/quizzes/{id}:
    get:
      description: Retorna métricas agregadas de um quiz.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Estatísticas do Quiz
      tags:
      - Reports
  /reports/rooms:
    get:
      description: Lista salas finalizadas do professor logado.
      parameters:
      - description: Página (default 1)
        in: query
        name: page
        type: integer
      - description: Limite (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/history.RoomHistory'
            type: array
      security:
      - BearerAuth: []
      summary: Histórico de Salas
      tags:
      - Reports
  /reports/rooms/{id}:
    get:
      description: Retorna detalhes completos de uma sala, incluindo ranking e stats.
      parameters:
      - description: History Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history.RoomHistory'
        "404":
          description: Sala não encontrada
      security:
      - BearerAuth: []
      summary: Detalhe da Sala
      tags:
      - Reports
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"